+++ db/migrate/20191017133711_allow_null_match_winner.rb
@@ -1,12 +1,12 @@
 class AllowNullMatchWinner < ActiveRecord::Migration[6.0]
   def up
-    change_column :matches, :winner_id, :integer, null: true
+    change_column :matches, :winner_id, :bigint, null: true
 
     add_index :players, :osu_id, unique: true, name: 'players_uniq_online_id'
   end
 
   def down
-    change_column :matches, :winner_id, :integer, null: false
+    change_column :matches, :winner_id, :bigint, null: false
 
     remove_index :players, name: 'players_uniq_online_id'
   end
